third_party/blink/web_tests/platform/generic/fast/forms/date/date-stepup-stepdown-from-renderer-expected.txt

Check if stepping-up and -down for date input fields from renderer does not work.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS stepUp("2010-02-10") is "2010-02-10"


PASS successfullyParsed is true

TEST COMPLETE
